Extract glossary from Platforms White Paper to top level concern
In writing the platforms white paper it was realised that glossary
would help clarify some key terms. These were originally added at
the end of the white paper for proximity and ease.

During the review of the upcoming platform engineering maturity
model it was realised that the entries in this glossary would be
useful to the maturity model readers as well.

This PR extracts the glossary into its own top level content. It
intentionally does not introduce any new terms or improve any
existing terms as that work should be evaluated in independent
PRs.

The only "decisions" made in this PR are:
1. The glossary deserves to be its own stand alone page
2. The glossary should be a TAG level concern (despite it currently
only containing WG Platform terms).

Signed-off-by: Abby Bangser <abby@syntasso.io>
Signed-off-by: Abby Bangser <abby@paintedwavelimited.com>