announcement blog and reformat of pdf
Our goal is to have this blog published by the CNCF but given the
tight timelines with KubeCon we will publish here first.

In addition, while creating the blog graphic I noticed a formatting
issue with the table where the last row was having some characters
cut off. I have widened the rows to remove this issue and re-
uploaded the PDF with only that change.

Signed-off-by: Abby Bangser <abby@syntasso.io>
Signed-off-by: Hoon Jo <pagaia@hotmail.com>