Introduce blog contribution standards
There is an intention to introduce more community blogs onto the TAG
website. This is being particularly pushed by the platforms WG where
it is an emergent space and we would like an opportunity to explore
some of the ideas with a lower barrier to entry. This will also enable
more inclusion as we will encourage and support new voices without
requiring 6-12 months of editing.

This page is meant to introduce a baseline requirement for all TAG
contributions. If this content is approved, it is expected that the
WG Platforms will release a blog which further extends this to match
the way that WG in particular would like to work with new authors.

Signed-off-by: Abby Bangser <abby@paintedwavelimited.com>