Contributions page (#519)
* Contributions page

We get a lot of questions about where to start with the TAG or a WG.
This page is a great location for us to start supporting people with
an easier path into the groups.

@techmaharaj and I put together a draft of something we could add
here to get started.

Signed-off-by: Abby Bangser <abby@paintedwavelimited.com>

* Introduce blog contribution standards

There is an intention to introduce more community blogs onto the TAG
website. This is being particularly pushed by the platforms WG where
it is an emergent space and we would like an opportunity to explore
some of the ideas with a lower barrier to entry. This will also enable
more inclusion as we will encourage and support new voices without
requiring 6-12 months of editing.

This page is meant to introduce a baseline requirement for all TAG
contributions. If this content is approved, it is expected that the
WG Platforms will release a blog which further extends this to match
the way that WG in particular would like to work with new authors.

Big thanks to @earth2marsh for putting together the key points for
this policy!

Signed-off-by: Abby Bangser <abby@paintedwavelimited.com>

---------

Signed-off-by: Abby Bangser <abby@paintedwavelimited.com>
Signed-off-by: Hoon Jo <pagaia@hotmail.com>