Introduce taxonomy for blog pages
Hugo has the built in ability to do categories and tags. Using them
can enable things like word clouds, pages that highlight aligned
content and more.

Right now this is just the first introduction of categories and tags.
This won't appear anywhere until additional styling is added to our
Hugo setup. This is being tracked in #522.

This commit just puts into place the categorisation/tagging for now.
The proposed taxonomy is:

* *Categories* should be minimal and applicable across a number of tags.
  Therefore there are only two:
  * announcements
  * article
* *Tags* can help identify sub topics that are more narrow focuses.
  Right now there are three proposed tags:
  * event
  * community contribtion
  * WG Multi-tenancy
  * WG Platforms

These will be codified in the issue template created under #495 and
possibly also on the contributions page.

Signed-off-by: Abby Bangser <abby@paintedwavelimited.com>