Introduce more guidance on getting start contributing (#523)
* Introduce more guidance on getting start contributing

We have made awesome strides by introducing more contributing
guides to the website in #519.

This closes #495 by introducing both a detailed blog post and
an accompanying issue template which can support organising
people's thoughts and provide the TAG with more information
to support new ideas.

This is heavily modelled after the sandbox template but of course
needs to be much more light weight. One thing that has been
removed but may need to be re-added is a reference to who
owns IP from any contribution.

Big thanks to @techmaharaj for a lot of the content and the
first review of this PR.

Signed-off-by: Abby Bangser <abby@paintedwavelimited.com>
Co-authored-by: Atulpriya Sharma <sharma.atulpriya@gmail.com>
Signed-off-by: Hoon Jo <pagaia@hotmail.com>