Introduce new templates
I attempted to introduce a template for new contributions but in using
the new format I added the wrong file extension. This is a format that
requires `.yml` not `.md` therefore the template is not visible on
the repo.

In addition, this commit also introduces another new template for
translations. This is in an attempt to make it easier for people to
add translations to the site. I have noticed that people are being
great about building off past PRs, but I thought codifying it in
a template could be helpful.

Signed-off-by: Abby Bangser <abby@paintedwavelimited.com>