Create section for "ever green" projects
In reviewing the comments on this PR, I realised that my attempt to organise
things had the very real risk of being too much process. With this in mind I
have removed the `platform maturity model examples` and moved it to a
description page within the contributions section on the website.

This led me to a small reorganisation of the contributions page and
clarification on when to use an issue (open and ongoing project) and when to
use this new page section (ever green, on-going opportunity).

Signed-off-by: Abby Bangser <abby@paintedwavelimited.com>