Introduce new templates (#525)
* Introduce new templates
I attempted to introduce a template for new contributions but in using
the new format I added the wrong file extension. This is a format that
requires `.yml` not `.md` therefore the template is not visible on
the repo.

In addition, this commit also introduces another new template for
translations. This is in an attempt to make it easier for people to
add translations to the site. I have noticed that people are being
great about building off past PRs, but I thought codifying it in
a template could be helpful.

* add context for the new issue templates

* Create section for "ever green" projects
In reviewing the comments on this PR, I realised that my attempt to organise
things had the very real risk of being too much process. With this in mind I
have removed the `platform maturity model examples` and moved it to a
description page within the contributions section on the website.

This led me to a small reorganisation of the contributions page and
clarification on when to use an issue (open and ongoing project) and when to
use this new page section (ever green, on-going opportunity).

---------
Signed-off-by: Abby Bangser <abby@paintedwavelimited.com>